extract screen to separate component
